MVP version. Need reformat code

+++ internal/domain/service.go
@@ -1,6 +1,6 @@
 package domain
 
-func (h *Holders) AddValue(address string, value int) {
+func (h *Holders) AddValue(address string, value float64) {
 	_, ok := h.Holders[address]
 	if !ok {
 		h.Holders[address] = value
@@ -9,7 +9,7 @@ func (h *Holders) AddValue(address string, value int) {
 	}
 }
 
-func (h *Holders) SubValue(address string, value int) {
+func (h *Holders) SubValue(address string, value float64) {
 	_, ok := h.Holders[address]
 	if !ok {
 		h.Holders[address] = value
