+++ internal/usecases/interfaces.go
@@ -1,5 +1,5 @@
 package usecases
 
 type UseCases interface {
-	FindMaxBalanceWalletForLastBlocks(numberBlocks int) (WalletBalanceModel, error)
+	FindMaxBalanceWalletForLastBlocks(numberBlocks int64) (WalletBalanceModel, error)
 }
